add statistiques

+++ src/main/java/com/comdata/backend/comdatapointage/dto/StatsParamDto.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.Date;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class StatsParamDto {
     private String date; //year or mounth or day //2020 or Janvier or 30
-    private Integer value;
+    private double value;
 }
